requirements.txt
Flask>=2.0.0
python-dotenv>=0.20.0
requests>=2.28.0
flask-cors>=3.0.0
